% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyncdf.R
\name{dimensions}
\alias{dimensions}
\title{Get the names and ranges of dimensions in a NetCDF file or connection.}
\usage{
dimensions(f)
}
\arguments{
\item{f}{either a string indicating a NetCDF file or a ncdf4 connection}
}
\description{
Get the names and ranges of dimensions in a NetCDF file or connection.
}
