% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyncdf.R
\docType{package}
\name{easyncdf}
\alias{easyncdf}
\alias{easyncdf-package}
\title{easyncdf is a collection of data processing verbs for working with NetCDF files.}
\description{
easyncdf is a collection of data processing verbs for working with NetCDF files.
}
\author{
Nicholas Potter
}
