% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyncdf.R
\name{read_ncdf}
\alias{read_ncdf}
\title{Read a ncdf4 file or connection and return a tidy data set.}
\usage{
read_ncdf(f, bounds, variables, ..., warn_cells = 1e+07, force = FALSE)
}
\arguments{
\item{f}{a NetCDF filename or connection.}

\item{bounds}{a named list of dimension bounds. Returned values are inclusive
of bounds.}

\item{variables}{a list of variable names to return.}

\item{...}{additional parameters passed to \code{nc_open()}.}

\item{warn_cells}{the number of returned cells (rows X columns) at which to
fail if \code{force} is not set to TRUE.}

\item{force}{force fetching data even if it will exceed 1 million records.}
}
\value{
a tidy data.frame.
}
\description{
Read a ncdf4 file or connection and return a tidy data set.
}
