% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyncdf.R
\name{summary.ncdf4}
\alias{summary.ncdf4}
\title{Summarize a ncdf4 file or connection.}
\usage{
\method{summary}{ncdf4}(f)
}
\arguments{
\item{f}{a NetCDF filename or connection.}
}
\value{
a table summarizing the NetCDF file.
}
\description{
Summarize a ncdf4 file or connection.
}
