% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyncdf.R
\name{variables}
\alias{variables}
\title{Get the names of variables in a NetCDF file or connection.}
\usage{
variables(f)
}
\arguments{
\item{f}{either a string indicating a NetCDF file or a ncdf4 connection}
}
\description{
Get the names of variables in a NetCDF file or connection.
}
